+++ spec/Receiver/SameBusReceiverSpec.php
@@ -2,8 +2,8 @@
 
 namespace spec\League\Tactician\Bernard\Receiver;
 
+use Bernard\Message;
 use League\Tactician\CommandBus;
-use League\Tactician\Bernard\QueueableCommand;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 
@@ -20,14 +20,14 @@ class SameBusReceiverSpec extends ObjectBehavior
         $this->shouldHaveType('League\Tactician\Bernard\Receiver');
     }
 
-    function it_handles_a_message(CommandBus $commandBus, QueueableCommand $command)
+    function it_handles_a_message(CommandBus $commandBus, Message $command)
     {
         $commandBus->handle(Argument::type('League\Tactician\Bernard\QueuedCommand'))->willReturn(true);
 
         $this->handle($command)->shouldReturn(true);
     }
 
-    function it_is_invokable(CommandBus $commandBus, QueueableCommand $command)
+    function it_is_invokable(CommandBus $commandBus, Message $command)
     {
         $commandBus->handle(Argument::type('League\Tactician\Bernard\QueuedCommand'))->willReturn(true);