+++ spec/Receiver/SeparateBusReceiverSpec.php
@@ -2,8 +2,8 @@
 
 namespace spec\League\Tactician\Bernard\Receiver;
 
+use Bernard\Message;
 use League\Tactician\CommandBus;
-use League\Tactician\Bernard\QueueableCommand;
 use PhpSpec\ObjectBehavior;
 
 class SeparateBusReceiverSpec extends ObjectBehavior
@@ -19,14 +19,14 @@ class SeparateBusReceiverSpec extends ObjectBehavior
         $this->shouldHaveType('League\Tactician\Bernard\Receiver');
     }
 
-    function it_handles_a_message(CommandBus $commandBus, QueueableCommand $command)
+    function it_handles_a_message(CommandBus $commandBus, Message $command)
     {
         $commandBus->handle($command)->willReturn(true);
 
         $this->handle($command)->shouldReturn(true);
     }
 
-    function it_is_invokable(CommandBus $commandBus, QueueableCommand $command)
+    function it_is_invokable(CommandBus $commandBus, Message $command)
     {
         $commandBus->handle($command)->willReturn(true);