+++ src/QueuedCommand.php
@@ -2,6 +2,8 @@
 
 namespace League\Tactician\Bernard;
 
+use Bernard\Message;
+
 /**
  * Indicates the command has been queued or not
  *
@@ -10,14 +12,14 @@ namespace League\Tactician\Bernard;
 class QueuedCommand
 {
     /**
-     * @var QueueableCommand
+     * @var Message
      */
     protected $command;
 
     /**
-     * @param QueueableCommand $command
+     * @param Message $command
      */
-    public function __construct(QueueableCommand $command)
+    public function __construct(Message $command)
     {
         $this->command = $command;
     }
@@ -25,7 +27,7 @@ class QueuedCommand
     /**
      * Returns the wrapped command
      *
-     * @return QueueableCommand
+     * @return Message
      */
     public function getCommand()
     {