+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Change Log
 
+
 ## 0.3.0 - 2015-03-30
 
 ### Changed
 
-- The `Command` interface from Tactician is no longer required.
+- The `Command` interface from Tactician is no longer required
+
+### Removed
+
+- `QueueableCommand` interface is not needed anymore
+
 
 ## 0.2.0 - 2015-03-17